Feat : 유튜브 검색 재생시간 수정중

+++ zizon/src/main/java/com/b302/zizon/domain/music/dto/response/MusicInfoResponseDTO.java
@@ -12,12 +12,13 @@ public class MusicInfoResponseDTO {
     private Long musicId;
     private String title;
     private String artist;
-    private int duration;
+    private Long duration;
     private String albumCoverUrl;
     private String youtubeVideoId;
 
     @Builder
-    public MusicInfoResponseDTO(Long musicId, String title, String artist, int duration, String albumCoverUrl, String youtubeVideoId) {
+
+    public MusicInfoResponseDTO(Long musicId, String title, String artist, Long duration, String albumCoverUrl, String youtubeVideoId) {
         this.musicId = musicId;
         this.title = title;
         this.artist = artist;
